architecture-spring-core/src/main/java/com/acme/architecture/core/client/dummy/DummyHeaderRequestInterceptor.java
package com.acme.architecture.core.client.dummy;

import java.io.IOException;

import org.springframework.http.HttpRequest;
import org.springframework.http.client.ClientHttpRequestExecution;
import org.springframework.http.client.ClientHttpRequestInterceptor;
import org.springframework.http.client.ClientHttpResponse;

public class DummyHeaderRequestInterceptor implements ClientHttpRequestInterceptor {

	@Override
	public ClientHttpResponse intercept(HttpRequest request, byte[] body, ClientHttpRequestExecution execution)
			throws IOException {
		return DummyClientHttpResponse.nextDummyResponse();
	}

}
